requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
greenlet==1.1.1
idna==3.2
importlib-metadata==4.7.1
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.23
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0
